+++ HISTORY.rst
@@ -12,6 +12,7 @@ End-User Summary
 - MutationTaster scoring now able to score InDels.
 - MutationTaster rank now displayed as numbers, not as stars, with -1 corresponding to an error during scoring.
 - Adding "closed uncertain" state.
+- Project-wide filtration allows for comp het filter for individual families.
 
 Full Change List
 ================
@@ -21,6 +22,7 @@ Full Change List
   Rank -1 and probability -1 correspond to error during MutationTaster ranking or empty results from MutationTaster.
 - Improving display and logging in alignment QC import.
 - Adding "closed uncertain" state.
+- Project-wide filtration allows for comp het filter for individual families.
 
 -------
 v0.17.5